requirements.txt
streamlit
pandas
google-genai
gTTS
python-dotenv